test: consolidate coverage suites
